output window: added allocation setting for drawing areas. fixes bug #45229.
I added allocation setting to the drawing areas in the output window. The
problem occured in osx gtk3 branch and is described in bug #45229. The
content of the output window was not rendered correctly in the initial
window but only after a resize. Setting the size request for the drawing
areas is not enough because the current allocation is not immediatly changed.
The clipping is based on the current allocation, while the drawing is based
on the results from the rerender function. This problem is related to
the problem described in bug #43362.