psppire-output-view.c: added check for view->xr - fixes bug #45421
The output windows crashed when an empty syntax window is executed. The
rerender function got called without a valid view->xr (=NULL). I added a
check for this condition and create a new view->xr as it is done in the
psppire_output_view_put function.

This fixes bug #45421: Crash when trying to run empty syntax