missing-val-dialog.c: fixed bug #45419 - added value type checks
This patch fixes bug #45419:
psppire crashes when entering missing data in new string variable.

The missing value width is now derived from the variable type. I noticed that
there are currently no type checks and no detailed warnings. I added the type
checks which were already provided by data_in and forward them to the existing
error dialog box. If an error is detected, the focus of the dialog box is
set to the field which produces the error.