dictionary: Make dict_delete_var() much faster in common case.
When a system file is opened, dict_delete_consecutive_vars() deletes lots
of variables if the system file has long string variables.  This was very
slow because of the O(n**2) behavior in dict_delete_var().  The biggest
cost of that O(n**2) behavior was in calling var_clone() on every variable
several times.  Fortunately, the var_clone() call is not necessary in this
case, because the cloned variable is only needed for callbacks, which
aren't set when a dictionary is being read.

This reduced the time to read the dictionary in a particular sample system
file from seconds to milliseconds.