Restore the splash screen.
Re-implementat the splash screen, using functinos which are not
deprecated under Gtk3.  Note:  This implementation does not
use the gtk functions which the Gtk devs recommend for splash screens.
What they forget, is that a primary purpose of a splash screen is
that it should appear with the minimum of delay.  gtk_init can
have significant delay - hence the recommended way of doing splash
screens in Gtk is self defeating.