Don't bother freeing import dialog memory. Thus avoid crashes.
A number of users have complained that the import dialog assistant
crashes on exit.  This seems to be because some callback is accessing
the data after it has been freed.

This change "fixes" that by simply not ever freeing the data.  This
will cause a memory leak.  However during my investigations, I have
found that this dialog leaks so much anyway that these few bytes are
hardly significant.

A proper fix will hopefully be forthcoming.

Affects bug #45805