Fix dereference to reallocated pointer in data-parser.c (parse_delimited_no_span, parse_delimited_span)
The parse_delimited_span and parse_delimited_no_span functions were
accessing a pointer from within a loop, but the pointer had been
initialised outside the loop.  However the loop itself was deallocating
the memory referenced by that pointer.

This change initialises the pointer within the loop so as to avoid
a potential crash here.