Make fn_open and fn_close take a struct file_handle instead of char *
This change alters the signature of the fn_open and fn_close functions,
and changes all callers appropriately.  The purpose of such a change is
so that the encoding of the filename is passed to fn_open from wherever
it was created.

It also changes the implementation of fn_open on windows systems to call
_wfopen instead of fopen.  This should fix the problem of windows builds
not being able to open files whose pathnames included non-ascii characters.

As a bonus, we no longer require the rather contrived function
local_to_filename_encoding.

Some related issues (what about the pspprc file?) remain.