Output renderer:  Reverse the order of table columns when in RTL locale
When operating in a RTL locale, the last column of a table should be rendered
"first" (ie. leftmost).  This change achieves that.  In order to determine if
the locale is a RTL locale, a new translatable string is introduced.  This
string must be localised by the translator of the .po file for RTL locales. LTR
locales can leave it untranslated.

Reported by: Mohammad Haghighat <hosseinhg@hotmail.com>