REGRESSION: Correctly handle data in temporary file for SAVE subcommand.
The format of the data in the temporary file used by the SAVE subcommand is
basically unrelated to the format of the data in the active file, but the
code that created and wrote cases to it created them as clones of cases
in the active file.  This is simply incorrect behavior and causes nasty
failures when the active file has an inconvenient number of variables or
contains long string variables.

This commit fixes the problem by correctly creating the cases for the
temporary file using that file's own caseproto.

In addition, the assignment of caseproto indexes to variables in the
temporary file was broken for /SAVE=PRED, because it would skip index 0.
This commit also fixes that problem.

Bug #44877.
Reported by Scott.