spreadsheet: Avoid sharing a dictionary between spreadsheet and client.
Spreadsheet implementations keep a copy of the dictionary that they build
to simplify reading cases.  Until now, they shared that dictionary, in a
modifiable way, with their client.  This meant, however, that if the client
modified the dictionary, it could screw up reading further cases from the
spreadsheet.

This commit fixes the problem by forcing clients to make their own copy of
the dictionary if they want to modify it.

Bug #44158.