histogram - changed bin range computation
The bin range computation for the histogram showed bin limits at 4.999999
when 5.0 was expected. Added values with value 5.0 were therefore added
to the following bin. With this change the bin limits do not show the
rounding error. At least I could not reproduce the error that showed
up when using the routing from gsl_histogram_set_ranges_uniform.