histogram - added another upper bin to include cases which have max values
This fixes the problem that cases which have values exactly at the upper
bin limit of the upper bin are not included in the binning. The fix
is based on the idea from John to just add another bin when the bin
limit is just equal to the maximum case value.