expressions: Merge all the little generator programs into generate.pl.
The little generator programs had a bizarre flaw: they silently produced
no output if there was an error parsing their input.  This was due to the
use of the Perl "do" construct for executing generate.pl, which just
returns the message if the nested code exits with "die".  I know that the
generators actually worked better at one point...

Anyway, the simplest solution seems to be to just merge all the little
output generator programs into generate.pl.  It doesn't make the code any
harder to understand, so that's what this commit does.