Remove attempts do destroy windows before quit.
This change deletes an attempt to remove all existing
windows from a list (and thereby unreffing and deleting them).
The attempt fails, because lists may not be altered whilst
iterating them.  This resulted only in a Critical Warning.

It is also pointless to attempt to destroy windows immediately
priot to quit.