adding osx application bundler scripts
This adds a utility script which produces a mac application bundle
based on the macports installation of pspp. The application bundle
can then be distributed as a single file and can be installed exactly
like a usual mac application. This is easier for mac users as they
do not need to install XCode, macports, XQuartz but just this package.