help-menu.c: fix default html browser start on windows - correct uri
commit #6c9aaa introduced calling the default html browser in case the
yelp system is not available on the system. This patch fixes a
problem on WINDOWS, where the browser did not open. The problem is
that on windows the "start" command is not a regular executable but
a part of the cmd shell. So the executable that has to be called is
"cmd" with start as a parameter to execute. There is an additional
problem on windows that the "Program Files" directory contains a
space in the path name. For a proper uri, this has to be
translated to Program%20Files. This is done via the
g_filename_to_uri function.