Update the Gnulib version
The primary motive for doing this is to update gl/version_etc which changes
the copyright year to 2016.  This is the value which gets displayed by --version
I do not think it is ideal that we must update gnulib to get the current year
in our version string, but I did not design this aspect of the code.