perl-module: Avoid VERSION redefinition warning from compiler.
If the PSPP version includes a suffix like "-pre1", then Perl will strip
that suffix from the version it passes to the compiler using -DVERSION=...,
and <config.h> will re-define it with the full version string.  This causes
a warning.

This change avoids the warning.  I am not yet aware of a new problem that
it causes.