output: Add support for cell margins.
Until now, space between cells was implemented by adding some space on the
sides of rules.  That worked fine before cells could have colored
backgrounds, but now it means that the backgrounds don't reach all the way
to the rules.  Thus, this commit adds configurable margins to the cells
themselves.