ascii: Get rid of the idea of pages.
In the era when I started writing this code, people still printed output on
dot-matrix printers and fanfold paper.  That's no longer a realistic use
case: today, no one prints plain text output.  It's just for viewing on a
screen, and on a screen there is no need or desire to break tables up
vertically into virtual "pages".