str: Fix bugs in ds_put_c_vformat().
This code used c_vasnprintf() as if it had the same interface as
vsnprintf(), but it's quite different.  It is necessary to make note of its
return value and possibly free it.  The code previously here didn't do
that, which led to memory leaks and possibly worse problems.  (Fortunately
this function isn't used much.)

Memory leak found by Address Sanitizer.