README.Git: Advance to Gnulib commit 9ce9be2ef0cb ("autoupdate").
A recent commit to PSPP depended on a more recent version of Gnulib that
includes Gnulib commit 3cd207e1d0fd ("fatal-signal: Pass the signal number
to the action.") but the PSPP commit did not actually advance the Gnulib
dependency to that commit.  This commit does that (actually it advances to
Gnulib tip of master).

Fixes: 21d2449f801c ("unlink_replace_files: Change signature to avoid compiler warning.")