LOOP: Limit number of iterations when IF clauses present.
The LOOP command is limited by the MXLOOPS setting in some cirumstances.
Until now, if an IF clause was present, MXLOOPS was disregarded.
However, this behavior does not match SPSS behavior.  This commit makes
LOOP honor MXLOOPS even when IF is present, for compatibility.

With this commit, only the presence of an index clause makes LOOP
disregard MXLOOPS.

Thanks to Frans Houweling for reporting this bug.