pivot-table: Kluge dropping default 9-point font to make output readable.
The default font size for SPV output is normally 9 points, but that
makes the output window hard to read.  This drops that default entirely,
which seems to make the output use a system or Pango default.  This is
probably not the right thing to do--probably we need a way to magnify
the fonts for on-screen display.

This reverts part of commit e3c3fbe8b, which wasn't intended to have a
visible effect but did because it inadvertently introduced a 9-point
default.

Reported by John Darrington.