cairo: Fix font scaling.
It's really a mystery why a 9-point font doesn't appear 9 points tall
in PDF (etc.) output, but it doesn't.  I guess we're doing something
wrong somewhere.  But until we figure out the underlying issue, we need
to scale down font sizes for PDF and not for on-screen use.

This is a slightly better fix for font sizes than the previous commit
3c15b41c036 (pivot-table: Kluge dropping default 9-point font to make
output readable.)