Makefile.am - added check for debian package directory
The debian package is now also maintained via git

https://salsa.debian.org/science-team/pspp

but that repository contains the distribution package. Due to the
.git directory the make process assumes that we are building
a distribution file and adds a check for some missing files. That
check fails if the distribution is under git control. I checked
for the existance of the debian directory as an indicator that
we are building the debian package and disable the check.