src/output/cairo-chart.h (xrchart_write_scale): Check the range and deal with the consequences.
If a chart was requested where either scale had an range of zero
(the minimum value and the maximum value were equal), then all
sorts of memory errors could occur later in the chart's rendering.

This change checks for this condition early and refuses to draw the
chart in this case.