fixed range check in parse_record_placement (32Bit/64Bit issue) #58968
The regression failed for test "lexer crash due to overflow" (403)
on i386 architecture. The reason is that long is 64Bit on amd64 while
it is 32Bit on i386. The code used long as type that can hold larger
values than int which does not work on 32Bit systems.

The change keeps the value as double as long as the range checks are done.

see: https://savannah.gnu.org/bugs/?58968