dummy.c: Conform to exact prototypes
Apparently certain toolchains complain if the symbols in dummy.c do not
exactly match their declarations.  This should not be a problem unless
the compiler has enabled a lot of warnings, and is treating those warnings
as errors, (which is inadvisable except under very special circumstances).
However, it appears that some people are doing that.

This change ensures that they are identical.