make-file: Pass O_BINARY or O_TEXT to open() call in replace_file_start().
It's hard to imagine why this would make a difference on a real system,
because the file descriptor obtained from open() is passed to fdopen(),
which is documented to determine whether the stream is text or binary
based on the mode string.  However, bug #40223 implies that it makes a
difference.  I hope that this commit fixes the problem (if any!).

Thanks to the anonymous submitter of bug #40223.