po/automake.mk: Fixed typo in variable names.
This resulted in po/pspp.pot not getting updated when it should have been.
This is a very good demonstration of the reason why carefully desigined
languages should insist on a variable being declared before it is used.

Also make the file depend on Makefile, so that it gets rebuilt when the
PACKAGE VERSION changes.