psppire-output-view: modified the way how copy and paste works
Copy (no Paste in output window) works now the usual gui way. You
can select an item in the output window like a table or a chart and
it will be highlighted. When you now do the copy operation via menu
or key, then the item is copied to the clipboard. With the change
the clipboard now provides the additional formats svg and various image
formats. You can now paste to inkscape, libreoffice suite and gimp.