doc: Improve formatting of tutorial in PDF output.
In the PDF version of the manual, excerpts from the output were formatted
awkwardly because output couldn't be split across pages.  In one case,
the output spilled off the bottom of the page.  This fixes the problem.

This change is pretty big.  It's because @float can't split across pages,
but without @float one must write in a different style that doesn't assume
the reader can jump back and forth to figures and tables.  I also took
the opportunity to start using pspp-output to excerpt portions of output.