MacOS: wc does not support -L option
The default wc application on MacOS does not support the -L option. I added
autotools support to search for GNU wc and use that instead. This assumes
that the application name is gwc which is the case whan using macports.

The tex tests use wc with the -L option. These regression tests fail on MacOS if the default wc application is used.