src/libpspp/Makefile.am: Fix atomicity of version.c generation.
If the Makefile has a bug, or someone runs two "make"s in parallel, the
approach of truncating version.c then appending to it line by line,
reopening the file each time, can wind up with a nasty partial result.  By
generating it in a single go, and then replacing the file atomically, we
avoid that problem.