psppire-output-view: Consistently apply "transparent" and "systemcolors".
Commit 81dece3642da ("output view: make items selectable and use system
colours") add transparency and system colors to the driver in the GUI,
but it did so somewhat after the driver was created.  I don't know why.
It makes more sense to me to set these options when creating the driver.
This commit makes that change.  It also deletes the function to change
the options; it had no other user.