psppire-window: Fix use-after-free error in read_spv_file().
spv_item_get_table() returns a borrowed reference, but read_spv_file()
treated it as if it owned it.  This fixes the problem.

This fixes crashes opening .spv files in PSPPIRE with File|Open.

Also, change spv_item_get_table() to return a const pointer, to make it
clearer that the reference is a borrowed one.