output: Replace OUTPUT_ITEM_PAGE_SETUP by a new driver function.
OUTPUT_ITEM_PAGE_SETUP was a special case that only the Cairo and SPV
drivers could implement, and it could only come first in the output stream
anyway, so it didn't make much sense.  This commit gets rid of it in favor
of adding a new driver class function for setting the page setup.