segment: Properly handle DO REPEAT in batch mode.
In batch mode, a line that begins with a non-blank starts a new command,
but the segmenter didn't properly support this.  Without this commit,
the following wasn't properly handled:

DO REPEAT #x=1
ECHO 'hi'
END REPEAT

This commit also adds a test against regression, which also checks that
a blank line to separate DO REPEAT from the repeated syntax works (it
did before too, but there was no test).