gui: For text data import, use the same parser for preview as for import.
There were odd discrepancies between the preview code in the GUI for
text data import and the code in GET DATA /TYPE=TXT that implemented the
eventual import.  This commit unifies the two parsers, dropping the one
that was used only for preview.  This should prevent further discrepancies.

Bug #61809.
Thanks to Friedrich Beckmann for reporting this bug.