perl-module: Copy Perl module from srcdir to builddir before distributing.
I ran into an issue where the Perl tests were failing in distcheck, but
not in check.  I changed the sources for the Perl module, but distcheck
still used the old ones.  It turns out that "make dist" uses files from
the builddir, which "make dist" wasn't updating from the srcdir because
that only happened when the Perl module was actually built, which "make
dist" and "make distcheck" don't do.

This commit fixes the problem by ensuring that the Perl module files get
updated in the builddir on "make dist".