Smake: Fix dependencies for src/ui/gui/resources.xml.
This file is a list of all the icon files, but generating it doesn't use
icon files themselves, so depending on them isn't useful.  On the other
hand, the file does need to be regenerated if the icon list changes, and
that icon list is in Smake itself, so this target does need to depend on
the Makefiles.