CTABLES: Make ID values reproducible when category labels are moved.
ID values were assigned in an arbitrary order when category labels were
moved, because the pivot table leaf index for the category dimension was
not included in the sort.  This fixes the problem.

This also factors finding the category leaf index out of the inner loop in
ctables_table_output().  This index doesn't vary across that inner loop and
I simply hadn't noticed while writing the code.